requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
aiosqlite
httpx_oauth
python-dotenv